+++ src/main/java/org/lndroid/framework/dao/ISubscribeSendPaymentsDao.java
@@ -6,5 +6,5 @@ import org.lndroid.framework.WalletData;
 
 public interface ISubscribeSendPaymentsDao {
     WalletData.SendPayment getPayment(long id);
-    List<WalletData.SendPayment> getActivePayments(int userId);
+    List<WalletData.SendPayment> getActivePayments(long userId);
 }
