+++ src/main/java/org/lndroid/framework/engine/IPluginBackgroundCallback.java
@@ -7,7 +7,7 @@ public interface IPluginBackgroundCallback {
     // not notified immediately, and thus
     // plugin can finish writing it's request before
     // clients start reading it.
-    int onAuthBackground(String pluginId, String type);
+    long onAuthBackground(String pluginId, String type);
 
     // notify other plugins about some change
     void onSignal(String pluginId, String topic, Object data);
