+++ src/main/java/org/lndroid/framework/room/ConnectPeerDao.java
@@ -25,7 +25,7 @@ abstract class ConnectPeerDaoRoom
     public abstract List<RoomTransactions.ConnectPeerTransaction> getTransactions();
 
     @Query("SELECT * FROM ConnectPeerTransaction WHERE txUserId = :txUserId AND txId = :txId")
-    public abstract RoomTransactions.ConnectPeerTransaction getTransaction(int txUserId, String txId);
+    public abstract RoomTransactions.ConnectPeerTransaction getTransaction(long txUserId, String txId);
 
     @Insert
     public abstract void createTransaction(RoomTransactions.ConnectPeerTransaction tx);
@@ -35,26 +35,26 @@ abstract class ConnectPeerDaoRoom
     @Query("UPDATE ConnectPeerTransaction " +
             "SET txAuthTime = :time, txAuthUserId = :txAuthUserId " +
             "WHERE txUserId = :txUserId AND txId = :txId")
-    public abstract void confirmTransaction(int txUserId, String txId, int txAuthUserId, long time);
+    public abstract void confirmTransaction(long txUserId, String txId, long txAuthUserId, long time);
 
     @Query("UPDATE ConnectPeerTransaction " +
             "SET txState = :txState, txDoneTime = :time, txAuthTime = :time, txAuthUserId = :txAuthUserId " +
             "WHERE txUserId = :txUserId AND txId = :txId")
-    public abstract void rejectTransaction(int txUserId, String txId, int txAuthUserId, int txState, long time);
+    public abstract void rejectTransaction(long txUserId, String txId, long txAuthUserId, int txState, long time);
 
     @Override
     @Query("UPDATE ConnectPeerTransaction " +
             "SET txState = :txState, txDoneTime = :time, txError = :code " +
             "WHERE txUserId = :txUserId AND txId = :txId")
-    public abstract void failTransaction(int txUserId, String txId, String code, int txState, long time);
+    public abstract void failTransaction(long txUserId, String txId, String code, int txState, long time);
 
     @Query("UPDATE ConnectPeerTransaction " +
             "SET txState = :txState, txDoneTime = :time " +
             "WHERE txUserId = :txUserId AND txId = :txId")
-    public abstract void timeoutTransaction(int txUserId, String txId, int txState, long time);
+    public abstract void timeoutTransaction(long txUserId, String txId, int txState, long time);
 
     @Override @Transaction
-    public void confirmTransaction(int txUserId, String txId, int txAuthUserId, long time, WalletData.ConnectPeerRequest authedRequest) {
+    public void confirmTransaction(long txUserId, String txId, long txAuthUserId, long time, WalletData.ConnectPeerRequest authedRequest) {
         if (authedRequest != null) {
             RoomTransactions.ConnectPeerTransaction tx = getTransaction(txUserId, txId);
             tx.txData.txAuthTime = time;
@@ -67,7 +67,7 @@ abstract class ConnectPeerDaoRoom
     }
 
     @Transaction
-    public WalletData.ConnectPeerResponse commitTransaction(int txUserId, String txId, WalletData.ConnectPeerResponse r, long time) {
+    public WalletData.ConnectPeerResponse commitTransaction(long txUserId, String txId, WalletData.ConnectPeerResponse r, long time) {
         RoomTransactions.ConnectPeerTransaction tx = getTransaction(txUserId, txId);
 
         // NOTE: r is not written to it's own table bcs we don't store addresses
