+++ src/main/java/org/lndroid/framework/room/IRoomActionDao.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 interface IRoomActionDao<Transaction, Response> {
     List<Transaction> getTransactions();
-    Transaction getTransaction(int txUserId, String txId);
+    Transaction getTransaction(long txUserId, String txId);
     void createTransaction(Transaction tx);
     void updateTransaction(Transaction tx);
-    void failTransaction(int txUserId, String txId, int txAuthUserId, int txState, long time);
-    Response commitTransaction(Transaction tx, int txAuthUserId, Response user, long time);
+    void failTransaction(long txUserId, String txId, long txAuthUserId, int txState, long time);
+    Response commitTransaction(Transaction tx, long txAuthUserId, Response user, long time);
 }
