+++ src/main/java/org/lndroid/framework/room/IRoomLndActionDao.java
@@ -4,12 +4,12 @@ import java.util.List;
 
 interface IRoomLndActionDao<Transaction, Request, Response> {
     List<Transaction> getTransactions();
-    Transaction getTransaction(int txUserId, String txId);
+    Transaction getTransaction(long txUserId, String txId);
     void createTransaction(Transaction tx);
     void updateTransaction(Transaction tx);
-    void confirmTransaction(int txUserId, String txId, int txAuthUserId, long time, Request authedRequest);
-    void rejectTransaction(int txUserId, String txId, int txAuthUserId, int txState, long time);
-    void failTransaction(int txUserId, String txId, String code, int txState, long time);
-    void timeoutTransaction(int txUserId, String txId, int txState, long time);
-    Response commitTransaction(int txUserId, String txId, Response r, long time);
+    void confirmTransaction(long txUserId, String txId, long txAuthUserId, long time, Request authedRequest);
+    void rejectTransaction(long txUserId, String txId, long txAuthUserId, int txState, long time);
+    void failTransaction(long txUserId, String txId, String code, int txState, long time);
+    void timeoutTransaction(long txUserId, String txId, int txState, long time);
+    Response commitTransaction(long txUserId, String txId, Response r, long time);
 }
